+++ src/app/commands/generate-express-controller.command.ts
@@ -30,6 +30,6 @@ export class GenerateExpressControllerCommand extends BaseCommand {
    * @param {Uri} folderPath - The folder path
    */
   async execute(folderPath?: Uri): Promise<void> {
-    this.service.generateComponent(folderPath, 'express-controller');
+    await this.service.generateComponent(folderPath, 'express-controller');
   }
 }