+++ src/app/commands/generate-express-middleware.command.ts
@@ -30,6 +30,6 @@ export class GenerateExpressMiddlewareCommand extends BaseCommand {
    * @param {Uri} folderPath - The folder path
    */
   async execute(folderPath?: Uri): Promise<void> {
-    this.service.generateComponent(folderPath, 'express-middleware');
+    await this.service.generateComponent(folderPath, 'express-middleware');
   }
 }