+++ src/app/commands/generate-express-server.command.ts
@@ -30,6 +30,6 @@ export class GenerateExpressServerCommand extends BaseCommand {
    * @param {Uri} folderPath - The folder path
    */
   async execute(folderPath?: Uri): Promise<void> {
-    this.service.generateComponent(folderPath, 'express-server');
+    await this.service.generateComponent(folderPath, 'express-server');
   }
 }