+++ src/app/commands/generate-fastify-middleware.command.ts
@@ -30,6 +30,6 @@ export class GenerateFastifyMiddlewareCommand extends BaseCommand {
    * @param {Uri} folderPath - The folder path
    */
   async execute(folderPath?: Uri): Promise<void> {
-    this.service.generateComponent(folderPath, 'class');
+    await this.service.generateComponent(folderPath, 'fastify-middleware');
   }
 }