+++ src/app/commands/generate-node-server.command.ts
@@ -30,6 +30,6 @@ export class GenerateNodeServerCommand extends BaseCommand {
    * @param {Uri} folderPath - The folder path
    */
   async execute(folderPath?: Uri): Promise<void> {
-    this.service.generateComponent(folderPath, 'node-server');
+    await this.service.generateComponent(folderPath, 'node-server');
   }
 }