+++ src/app/commands/generate-type.command.ts
@@ -2,17 +2,17 @@ import { Uri } from 'vscode';
 import { BaseCommand } from './base.command';
 
 /**
- * The GenerateInterfaceCommand class.
+ * The GenerateTypeCommand class.
  *
  * @class
- * @classdesc The class that represents the generate interface command.
+ * @classdesc The class that represents the generate type command.
  * @extends {BaseCommand}
  * @export
  * @public
  * @example
- * const command = new GenerateInterfaceCommand(config);
+ * const command = new GenerateTypeCommand(config);
  */
-export class GenerateInterfaceCommand extends BaseCommand {
+export class GenerateTypeCommand extends BaseCommand {
   // -----------------------------------------------------------------
   // Methods
   // -----------------------------------------------------------------
@@ -30,6 +30,6 @@ export class GenerateInterfaceCommand extends BaseCommand {
    * @param {Uri} folderPath - The folder path
    */
   async execute(folderPath?: Uri): Promise<void> {
-    this.service.generateComponent(folderPath, 'interface');
+    await this.service.generateComponent(folderPath, 'type');
   }
 }